moved template packages processing to separate class
